画面のキャンバスに関連する処理をImageCanvasに移動しました。
理由は、widgets.pyのコードが500行を超えていたのて。
