+++ docs/CONTRIBUTING.md
@@ -21,10 +21,8 @@
     - `./run.bat`：  app.pyを実行するスクリプト。またはVSCodeの実行ボタンより。  
     - `./build.bat`： 実行ファイルを作成するスクリプト。PyInstallerライブラリを使用して生成します。  
     - `./release.bat`： 配布用zipファイルを作成するスクリプト。release.pyを使用します。  
-- 🎨`/dist`: 実行ファイル生成フォルダ。  
-    視覚的な要素を提供します。  
-- 🎨`/dist`: 実行ファイル生成フォルダ。  
-    視覚的な要素を提供します。  
+- 🎨`/assets`: アプリで使用するプログラムアイコン  
+- 📦`/dist`: 実行ファイル生成フォルダ。配布物のディレクトリ  
 - 🎨`/third_party`:サードパーティのリソースを含むディレクトリ。   
   - `/third_party/icons`:アプリで使用するアイコン。  
 - 📚`/docs`: ドキュメンテーション。
